Refactor to Next.js API
