+++ .prettierrc
@@ -1,0 +1,7 @@
+{
+	"bracketSpacing": false,
+	"semi": false,
+	"singleQuote": true,
+	"useTabs": true,
+	"printWidth": 110
+}
