--- api/index.js
@@ -1,5 +1,0 @@
-module.exports = (req, res) => {
-    res.status(200).json({
-	urlEndpoint: '/embed'
-    })
-}
