+++ package.json
@@ -1,21 +1,23 @@
 {
-   "name": "radio4000-api",
-   "version": "0.0.1",
-   "description": "",
-   "main": "/api/index.js",
-   "dependencies": {
-      "dompurify": "^2.2.2",
-      "got": "^11.8.0",
-      "jsdom": "^16.4.0",
-      "radio4000-sdk": "^0.1.3"
-   },
-   "devDependencies": {
-      "vercel": "^20.1.2"
-   },
-   "scripts": {
-      "test": "echo \"Error: no test specified\" && exit 1",
-      "start": "vercel dev"
-   },
-   "author": "",
-   "license": "aGPLv3"
+	"name": "radio4000-api",
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint"
+	},
+	"dependencies": {
+		"next": "12.0.7",
+		"react": "17.0.2",
+		"react-dom": "17.0.2",
+		"dompurify": "^2.3.4",
+		"got": "^12.0.0",
+		"jsdom": "^19.0.0",
+		"radio4000-sdk": "^0.1.3"
+	},
+	"devDependencies": {
+		"eslint": "8.5.0",
+		"eslint-config-next": "12.0.7"
+	},
+	"license": "aGPLv3"
 }
