+++ pages/api/index.js
@@ -1,7 +1,8 @@
-const config = require('../config')
+import config from '../../utils/config'
 
-module.exports = (req, res) => {
-  res.json({
+// Next.js API route support: https://nextjs.org/docs/api-routes/introduction
+export default function handler(req, res) {
+	const welcome = {
 		message: 'Welcome to the Radio4000 api',
 		databaseDocumentationUrl: 'https://github.com/internet4000/radio4000-firebase-rules',
 		apiDocumentationUrl: 'https://github.com/internet4000/radio4000-api-vercel',
@@ -11,6 +12,7 @@ module.exports = (req, res) => {
 		tracksUrl: `${config.databaseURL}tracks.json`,
 		trackUrl: `${config.databaseURL}tracks/{id}.json`,
 		channelEmbedUrl: `${config.apiURL}/embed?slug={slug}`,
-		channelOEmbedUrl: `${config.apiURL}/oembed?slug={slug}`
-  })
+		channelOEmbedUrl: `${config.apiURL}/oembed?slug={slug}`,
+	}
+	res.status(200).json(welcome)
 }
