--- vercel.json
@@ -1,8 +1,0 @@
-{
-    "routes": [
-	{ "src": "/", "dest": "/api/homepage.js" },
-	{ "src": "/embed", "dest": "/api/embed/index.js" },
-	{ "src": "/oembed", "dest": "/api/oembed/index.js" },
-	{ "src": "/backup", "dest": "/api/backup/index.js" }
-    ]
-}
